Content Helper: Move ts/tsx/scss files
