+++ src/content-helper/common/css/variables.scss
