+++ src/content-helper/common/icons/edit-icon.tsx
