+++ src/content-helper/common/icons/leaf-icon.tsx
