+++ src/content-helper/common/utils/number.ts
