+++ src/content-helper/dashboard-widget/dashboard-widget.scss
@@ -1,5 +1,5 @@
-@import "../../css/shared/variables";
-@import "../../css/shared/functions";
+@import "../common/css/variables";
+@import "../common/css/functions";
 
 #wp-parsely-dashboard-widget {
