+++ src/content-helper/editor-sidebar/editor-sidebar.scss
@@ -1,5 +1,5 @@
-@import "../../css/shared/variables";
-@import "../../css/shared/functions";
+@import "../common/css/variables";
+@import "../common/css/functions";
 
 .wp-parsely-content-helper {
