+++ src/content-helper/editor-sidebar/performance-details/model.ts
