+++ tests/js/content-helper/structure.test.tsx
@@ -11,10 +11,10 @@ import '@testing-library/jest-dom';
 /**
  * Internal dependencies.
  */
-import RelatedTopPostList from '../../../src/blocks/content-helper/related-top-posts/component-list';
-import RelatedTopPostsProvider, { GetRelatedTopPostsResult, RELATED_POSTS_DEFAULT_LIMIT, RELATED_POSTS_DEFAULT_TIME_RANGE } from '../../../src/blocks/content-helper/related-top-posts/provider';
-import { DASHBOARD_BASE_URL } from '../../../src/blocks/shared/utils/constants';
-import { ContentHelperError, ContentHelperErrorCode } from '../../../src/blocks/content-helper/content-helper-error';
+import RelatedTopPostList from '../../../src/content-helper/editor-sidebar/related-top-posts/component-list';
+import RelatedTopPostsProvider, { GetRelatedTopPostsResult, RELATED_POSTS_DEFAULT_LIMIT, RELATED_POSTS_DEFAULT_TIME_RANGE } from '../../../src/content-helper/editor-sidebar/related-top-posts/provider';
+import { DASHBOARD_BASE_URL } from '../../../src/content-helper/common/utils/constants';
+import { ContentHelperError, ContentHelperErrorCode } from '../../../src/content-helper/common/content-helper-error';
 
 describe( 'PCH Editor Sidebar Related Top Post panel', () => {
 	test( 'should display spinner when starting', async () => {
